# swagger/doc/cms.json
{
	"swagger": "2.0",
	"info": {
		"title": "Cozrum",
		"version": "1",
		"description": "Response format: #definitions/Response"
	},
	"basePath": "/api/v1",
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"security": [
		{
			"JWT": []
		}
	],
	"paths": {
		"/user/auth": {
			"post": {
				"tags": [
					"user"
				],
				"operationId": "authenticate",
				"summary": "Đăng nhập",
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"username": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"user": {
									"$ref": "#/definitions/User"
								},
								"token": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/user": {
			"get": {
				"tags": [
					"user"
				],
				"operationId": "getUserInfo",
				"summary": "Lấy thông tin tài khoản",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"user": {
									"$ref": "#/definitions/User"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"user"
				],
				"operationId": "updateUserInfo",
				"summary": "cập nhật thông tin user",
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"user": {
									"$ref": "#/definitions/User"
								}
							}
						}
					}
				}
			}
		},
...
		"/user/changePassword": {
			"post": {
				"tags": [
					"user"
				],
				"operationId": "changePassword",
				"summary": "Đổi mật khẩu",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Account",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"oldPassword": {
									"type": "string",
									"default": "admin"
								},
								"newPassword": {
									"type": "string",
									"default": "admin"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
		"/user/noti": {
			"put": {
				"tags": [
					"user"
				],
				"operationId": "updateNotiKey",
				"summary": "Cập nhật notify key",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "deviceName",
						"description": "ios/android/web",
						"required": true,
						"type": "string",
						"enum": [
							"ios",
							"android",
							"web"
						]
					},
					{
						"in": "query",
						"name": "notiId",
						"description": "notiId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
		"/user/calling/{userId}": {
			"put": {
				"tags": [
					"user"
				],
				"operationId": "updateCallingUser",
				"summary": "Cập nhật user gọi điện",
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "ServiceStringeeUser",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"order": {
									"type": "object",
									"properties": {
										"monday": {
											"type": "number"
										},
										"tuesday": {
											"type": "number"
										},
										"wednesday": {
											"type": "number"
										},
										"thursday": {
											"type": "number"
										},
										"friday": {
											"type": "number"
										},
										"saturday": {
											"type": "number"
										},
										"sunday": {
											"type": "number"
										}
									}
								},
								"days": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"time_start": {
												"type": "string"
											},
											"time_end": {
												"type": "string"
											},
											"disabled_calling": {
												"type": "boolean"
											},
											"disabled_listening": {
												"type": "boolean"
											},
											"day": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
		"/note": {
			"get": {
				"tags": [
					"note"
				],
				"operationId": "getNote",
				"summary": "Lấy Notes",
				"parameters": [
					{
						"in": "query",
						"name": "blockId",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "roomId",
						"type": "string"
					},
					{
						"in": "query",
						"name": "from",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "ok"
					}
				}
			},
			"post": {
				"tags": [
					"note"
				],
				"operationId": "addNotes",
				"summary": "add Notes",
				"parameters": [
					{
						"in": "query",
						"name": "blockId",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "roomId",
						"type": "string"
					},
					{
						"in": "query",
						"name": "date",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"notes": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/History"
									}
								},
								"rooms": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"schema": {
							"type": "object",
							"properties": {
								"ota": {
									"$ref": "#/definitions/OTAConfig"
								}
							}
						}
					}
				}
			}
		},
		"/note/{noteId}": {
			"put": {
				"tags": [
					"note"
				],
				"operationId": "updateNote",
				"summary": "update Note",
				"parameters": [
					{
						"in": "path",
						"name": "noteId",
						"description": "noteId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "note",
						"required": true,
						"schema": {
							"type": "object",
							"$ref": "#/definitions/History"
						}
					}
				],
				"responses": {
					"200": {
						"description": "ok"
					}
				}
			},
			"delete": {
				"tags": [
					"note"
				],
				"operationId": "removeNote",
				"summary": "remove Notes",
				"parameters": [
					{
						"in": "path",
						"name": "noteId",
						"description": "noteId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "ok"
					}
				}
			}
		},
		"/host/{hostID}": {
			"get": {
				"tags": [
					"host"
				],
				"operationId": "getHost",
				"summary": "Lấy thông tin Host",
				"parameters": [
					{
						"in": "path",
						"name": "hostID",
						"description": "Host ID",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"host": {
									"$ref": "#/definitions/Host"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"host"
				],
				"operationId": "updateHost",
				"summary": "Cập nhật thông tin Host",
				"parameters": [
					{
						"in": "path",
						"name": "hostID",
						"description": "Host ID",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Host model",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Host"
						}
					}
				],
				"responses": {
					"200": {
						"description": "host:",
						"schema": {
							"type": "object",
							"properties": {
								"host": {
									"$ref": "#/definitions/Host"
								}
							}
						}
					}
				}
			}
		},
...
		"/room/block/reorder": {
			"post": {
				"tags": [
					"room"
				],
				"operationId": "reorder",
				"summary": "Sắp xếp thứ tự blocks",
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"blocks": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"_id": {
												"type": "string"
											},
											"order": {
												"type": "number"
											}
										}
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "success"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
		"/task/{taskId}/note": {
			"post": {
				"tags": [
					"task"
				],
				"operationId": "addNote",
				"summary": "Thêm ghi chú",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"description": "taskId",
						"required": false,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"note": {
									"type": "string"
								},
								"images": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/task/{taskId}/note/{noteId}": {
			"put": {
				"tags": [
					"task"
				],
				"operationId": "updateNote",
				"summary": "Cập nhật ghi chú",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"description": "taskId",
						"required": false,
						"type": "string"
					},
					{
						"in": "path",
						"name": "noteId",
						"description": "noteId",
						"required": false,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"note": {
									"type": "string"
								},
								"images": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
		"/workNote?skip={skip}&limit={limit}&from={from}&to={to}&status={status}&blockId={blockId}": {
			"get": {
				"tags": [
					"workNote"
				],
				"operationId": "getWorkNotes",
				"summary": "get workNotes",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "skip",
						"type": "number"
					},
					{
						"in": "path",
						"name": "limit",
						"type": "number"
					},
					{
						"in": "path",
						"name": "from",
						"type": "string"
					},
					{
						"in": "path",
						"name": "to",
						"type": "string"
					},
					{
						"in": "path",
						"name": "status",
						"type": "string",
						"enum": [
							"doing",
							"done",
							"closed"
						]
					},
					{
						"in": "path",
						"name": "blockId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/WorkNote"
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/workNote": {
			"post": {
				"tags": [
					"workNote"
				],
				"operationId": "createWorkNote",
				"summary": "create workNote",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/WorkNote"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/WorkNote"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/workNote/{id}": {
			"put": {
				"tags": [
					"workNote"
				],
				"operationId": "updateWorkNote",
				"summary": "update workNote",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/WorkNote"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/WorkNote"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
	},
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"in": "header",
			"name": "x-access-token"
		}
	},
	"definitions": {
		"User": {
			"type": "object",
			"properties": {
				"username": {
					"type": "string"
				},
				"passwords": {
					"type": "string"
				},
				"group": {
					"type": "string"
				},
				"enable": {
					"type": "boolean"
				},
				"name": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"address": {
					"type": "string"
				},
				"phone": {
					"type": "string"
				}
			}
		},
		"Host": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string"
				},
				"blockIds": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Block"
					}
				},
				"ottIds": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/OTT"
					}
				},
				"hosting": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Block"
					}
				}
			}
		},
...
		"Block": {
			"type": "object",
			"properties": {
				"info": {
					"type": "object",
					"properties": {
						"name": {
							"type": "string"
						},
						"description": {
							"type": "string"
						},
						"address": {
							"type": "string"
						},
						"city": {
							"type": "string"
						},
						"location": {
							"type": "object",
							"properties": {
								"coordinates": {
									"type": "array",
									"items": {
										"type": "number"
									}
								}
							}
						},
						"blog": {
							"type": "string"
						},
						"order": {
							"type": "number"
						}
					}
				},
				"virtualRoomId": {
					"type": "string"
				},
				"listingIds": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"OTAProperties": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"otaName": {
								"type": "string"
							},
							"propertyId": {
								"type": "string"
							}
						}
					}
				},
				"lock": {
					"url": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"ottId": {
					"type": "string"
				},
				"active": {
					"type": "boolean"
				},
				"isProperty": {
					"type": "boolean"
				},
				"isTest": {
					"type": "boolean"
				},
				"staffs": {
					"type": "object",
					"properties": {
						"maid": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"OTAConfig": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"active": {
					"type": "boolean"
				},
				"cookie": {
					"type": "string"
				},
				"token": {
					"type": "string"
				},
				"other": {
					"type": "object",
					"properties": {
						"anything": {
							"type": "string"
						}
					}
				},
				"account": {
					"type": "string"
				},
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			}
		},
...
		"History": {
			"type": "object",
			"properties": {
				"_id": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"images": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"createdAt": {
					"type": "string"
				},
				"system": {
					"type": "boolean"
				},
				"by": {
					"type": "object",
					"properties": {
						"_id": {
							"type": "string"
						},
						"username": {
							"type": "string"
						}
					}
				},
				"removedBy": {
					"type": "object",
					"properties": {
						"_id": {
							"type": "string"
						},
						"username": {
							"type": "string"
						}
					}
				}
			}
		},
		"Task": {
			"type": "object",
			"properties": {
				"createdBy": {
					"type": "string"
				},
				"assigned": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"blockId": {
					"type": "string"
				},
				"bookingId": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"roomIds": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"description": {
					"type": "string"
				},
				"time": {
					"type": "string"
				},
				"status": {
					"type": "string",
					"enum": [
						"waiting",
						"confirmed",
						"done",
						"deleted"
					]
				}
			}
		},
		"OTT": {
			"type": "object",
			"properties": {
				"phone": {
					"type": "string"
				}
			}
		},
...
		"WorkNote": {
			"type": "object",
			"properties": {
				"blockId": {
					"type": "string"
				},
				"date": {
					"type": "string"
				},
				"note": {
					"type": "string"
				},
				"createdBy": {
					"type": "string"
				},
				"status": {
					"type": "string",
					"enum": [
						"doing",
						"done",
						"closed"
					]
				},
				"images": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"log": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"user": {
								"type": "string"
							},
							"date": {
								"type": "string"
							},
							"oldData": {
								"type": "object"
							},
							"newData": {
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	}
}